requirements.txt
pandas==2.1.4
openpyxl==3.1.2
ics==0.7.2
python-dateutil==2.8.2 